requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
django-address==0.2.5
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-paypal==1.0.0
django-taggit==1.3.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==2.0.0
psycopg2==2.8.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
